Stop truncating a range subscript's length to 32 bits (#354)
The count a range names was held in an int, so a dimension longer
than 2**31 got a truncated one. Up to 2**32 the truncation came out
negative and the guard below rounded it to an empty view, which is
how a[true, 0...(2**31)] answered a shape of 0. Past 2**32 it came
out positive and wrong, and the subscript silently named the low 32
bits of what was asked for: a[true, 0...((1<<32)+5)] on a longer
dimension answered 5 elements.

The count is now an ssize_t. The guard stays for a range that runs
backwards.

+++ ext/cumo/narray/index.c
@@ -252,7 +252,11 @@ cumo_na_parse_narray_index(VALUE a, int orig_dim, ssize_t size, cumo_na_index_ar
 static void
 cumo_na_parse_range(VALUE range, ssize_t step, int orig_dim, ssize_t size, cumo_na_index_arg_t *q)
 {
-    int n;
+    // The count is what the subscript names, which an int cannot hold once the
+    // dimension is longer than 2**31. The guard below only looked like it
+    // caught that: a truncated count came out negative up to 2**32 and was
+    // rounded to an empty view, and past that it came out positive and wrong.
+    ssize_t n;
     ssize_t beg, end, beg_orig, end_orig;
     const char *dot = "..", *edot = "...";
